refactor: move pkg/extensions/search/common/oci_layout.go under pkg/test/
