+++ pkg/extensions/search/resolver.go
@@ -1213,67 +1213,3 @@ func getReferrers(repoDB repodb.RepoDB, repo string, referredDigest string, arti
 
 	return results, nil
 }
-
-// get passed context from authzHandler and filter out repos based on permissions.
-func userAvailableRepos(ctx context.Context, repoList []string) ([]string, error) {
-	var availableRepos []string
-
-	// authz request context (set in authz middleware)
-	acCtx, err := localCtx.GetAccessControlContext(ctx)
-	if err != nil {
-		err := zerr.ErrBadType
-
-		return []string{}, err
-	}
-
-	if acCtx != nil {
-		for _, r := range repoList {
-			if acCtx.IsAdmin || acCtx.CanReadRepo(r) {
-				availableRepos = append(availableRepos, r)
-			}
-		}
-	} else {
-		availableRepos = repoList
-	}
-
-	return availableRepos, nil
-}
-
-func extractImageDetails(
-	ctx context.Context,
-	layoutUtils common.OciLayoutUtils,
-	repo, tag string, //nolint:unparam // function only called in the tests
-	log log.Logger) (
-	godigest.Digest, *ispec.Manifest, *ispec.Image, error,
-) {
-	validRepoList, err := userAvailableRepos(ctx, []string{repo})
-	if err != nil {
-		log.Error().Err(err).Msg("unable to retrieve access token")
-
-		return "", nil, nil, err
-	}
-
-	if len(validRepoList) == 0 {
-		log.Error().Err(err).Msg("user is not authorized")
-
-		return "", nil, nil, zerr.ErrUnauthorizedAccess
-	}
-
-	manifest, dig, err := layoutUtils.GetImageManifest(repo, tag)
-	if err != nil {
-		log.Error().Err(err).Msg("Could not retrieve image ispec manifest")
-
-		return "", nil, nil, err
-	}
-
-	digest := dig
-
-	imageConfig, err := layoutUtils.GetImageConfigInfo(repo, digest)
-	if err != nil {
-		log.Error().Err(err).Msg("Could not retrieve image config")
-
-		return "", nil, nil, err
-	}
-
-	return digest, &manifest, &imageConfig, nil
-}
